Incorporating code review feedback. - Moving process simulation info service into its own module - adding two additional tests to encapsulate the rest to test out sonar cloud coverage
