+++ Apromore-Custom-Plugins/Process-Simulation-Info-Logic/build.gradle
@@ -1,0 +1,7 @@
+group = 'org.apromore'
+version = '2.0.0'
+description = 'Process Simulation Info Logic'
+
+dependencies {
+	implementation project(':Apromore-Custom-Plugins:Log-Logic')
+}
